requirements.txt
discord.py
ics
python-dotenv
